Prepare regeneration of object-store
Since some time object-store service was not regenerated since the
schema defaulted to base off the AUTH_PROJECT_ID (as current python
tooling does). This is all fine, but is making it very hard to allow
project switching and info querying. The schema is being updated
separately, but we need some changes in the templates so that generated
code work.

Change-Id: I11f3f267b45dffc0a39e2675ec333ec5e3c1f269