Add metadata
Having OpenAPI is not enough to render sdk/cli/etc. A metadata is
necessary describing path to the spec, operation definition in the spec
and what and how to generate out of it. While most of the data can be
"assumed" there are cases where customization is necessary.

For now add metadata into the tree.

Change-Id: Ice129db1ceb2487c95d4c66946c6a0f57c60824e