Prepare git patch and archive for rust package
Apply the generated sources on top of the target repository and create
patch file. In addition to that save complete content as job artifacts.

post job is pushing changes to the upstream repository (using public SSH
key of repository set by Zuul and added as a deployment key). PR is not
created, since that requires additionally separate API key.

Change-Id: I090b87e05ede95f8542a427207bea711e5cc5df4