Update neutron schema resolution
Neutron treats subnetpool.XXX_prefixlen everywhere as integers including
the database (it is even described like that in neutron api-ref itself).
For some unknown reason it returns it as a string though. Maybe because
in the apidef it is defined like `validate: "type:non_negative",
convert_to: convert_to_int`. Modify api_def processing not to look at
validate OR convert_to, but use convert_to results in update of the
schema.

Change-Id: I52d1d207be54967a0a1b9a3443c661b1fabd69f4