Add support for explicit empty schemas
There might be a decorator explicitly stating the schema is empty
meaning there is no body (this is different from schema with type
"null"). In order to differentiate the empty schema from not set add a
similar Sentinet `UNSET` as in OpenStackSDK and generate dummy schemas
only when the schema is really unknown.

Change-Id: I72434afea54af9aa6df205a27a8d66765bb0ffd7